break out LicenseSerializer
